% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{access_token}
\alias{access_token}
\title{Access Meltwater token}
\usage{
access_token()
}
\description{
access_token() generates a Bearer token that is saved to your .Renvironment file. It is valid for 1 hour. If you run out of time, just run the function again.
}
