% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_document_authors.R
\name{interval_document_authors}
\alias{interval_document_authors}
\title{Get top document authors over a time period. Only available for social posts.}
\usage{
interval_document_authors(start_date, end_date, keyword = NULL,
  search_id = NULL, granularity = "DAY")
}
\arguments{
\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{keyword}{the keyword you want to explore, e.g. "iphone"}

\item{search_id}{search ID}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are available.}
}
\description{
interval_document_authors() generates a data frame containing the authors of documents written about a keyword or search id during an interval of time
}
\examples{
df <- interval_document_authors(keyword = "demoskop", start_date = "2017-07-01", end_date = "2017-07-13", type = "news_article")
}
