% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_document_count.R
\name{interval_document_count}
\alias{interval_document_count}
\title{Get document counts over a time period}
\usage{
interval_document_count(start_date, end_date, keyword = NULL,
  search_id = NULL, granularity = "DAY", type = "news")
}
\arguments{
\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format'}

\item{keyword}{the keyword you want to explore, e.g. "iphone"}

\item{search_id}{search ID}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are possible}

\item{type}{defaults to NULL and returns "all", can be news" or "social"}
}
\description{
interval_document_count() generates a data frame containing the number of documents written about a keyword or search id during an interval of time
}
\examples{
df <- interval_document_count(start_date = "2017-07-01", end_date = "2017-07-13", keyword = "demoskop", type = "news")
}
