% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_source_name.R
\name{interval_source_name}
\alias{interval_source_name}
\title{Get top source names for social post over a time period. Only available for social posts.}
\usage{
interval_source_name(start_date, end_date, keyword = NULL, search_id = NULL,
  granularity = "DAY")
}
\arguments{
\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{keyword}{the keyword you want to explore, e.g. "iphone"}

\item{search_id}{an ID for a search in Meltwater}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are available.}
}
\description{
interval_source_name() generates a data frame containing the number of documents written about a keyword or search id during an interval of time
}
\examples{
df <- interval_source_name(start_date = "2017-07-01",
 end_date = "2017-07-13", keyword = "demoskop")
}
