% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_sum_reach.R
\name{interval_sum_reach}
\alias{interval_sum_reach}
\title{Search counts of keyword during time interval}
\usage{
interval_sum_reach(start_date, end_date, keyword = NULL, search_id = NULL,
  type = NULL, granularity = "DAY")
}
\arguments{
\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{keyword}{the keyword you want to explore, e.g. "iphone"}

\item{type}{defaults to NULL and returns "all", can be news" or "social"}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are possible.}
}
\description{
interval_sum_reach() generates a data frame with the sum of reach over a period for a keyword
}
\examples{
df <- interval_sum_reach(start_date = "2017-07-01", keyword = "demoskop", type = "news", granularity = "DAY")
}
