% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_searches.R
\name{list_searches}
\alias{list_searches}
\title{Get a list of all your searches in Meltwater}
\usage{
list_searches(include_query = "false")
}
\arguments{
\item{include_query}{include a query with information such as query type, case sensitivity etc.}
}
\description{
list_searches() generates a data frame containing all your searches in Meltwater with Search ID
}
\examples{
df <- list_searches()
}
