% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_client.R
\name{reg_client}
\alias{reg_client}
\title{Request client_id and client_secret}
\usage{
reg_client(user_key, username, password)
}
\arguments{
\item{user_key}{your user key generated from https://developer.meltwater.com}

\item{username}{your username, most likely your e-mail}

\item{password}{your Meltwater password}
}
\description{
reg_client() enables you to autheticate yourself to Meltwaters API, it generates your client_id and client_secret, you will only aquire these once so save them somewhere
}
