% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_documents.R
\name{search_documents}
\alias{search_documents}
\title{Search documents}
\usage{
search_documents(start_date, end_date, keyword = NULL, search_id = NULL,
  type = "news", page_size = "30")
}
\arguments{
\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{keyword}{the keyword you want to explore}

\item{search_id}{an ID for a search in Meltwater}

\item{type}{"news" or "social"}

\item{page_size}{number of rows in data frame, max is 1000.}
}
\description{
search_documents() enables you to search every document containg a keyword during a specific time interval
}
\examples{
df <- search_documents(start_date = "2017-07-01",
 end_date = "2017-07-13",keyword = "demoskop", type = "news")
}
