% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interval.R
\name{search_interval}
\alias{search_interval}
\title{Search counts of keyword during time interval}
\usage{
search_interval(user_key, token, keyword, start_date, end_date, type,
  granularity)
}
\arguments{
\item{user_key}{your user key generated from https://developer.meltwater.com}

\item{token}{your token generated from access_token()}

\item{keyword}{the keyword you want to explore}

\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{type}{"news_article", "social_post" or "all", defaults to "all"}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are possible.}
}
\description{
search_interval() generates a data frame containing the number of articles or social posts distributed by time during a certain interval of time
}
