% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{access_token}
\alias{access_token}
\title{Access Meltwater token}
\usage{
access_token(client_id, client_secret, user_key)
}
\arguments{
\item{client_id}{your id, generated from reg_client()}

\item{client_secret}{your secret, generated from reg_client}

\item{user_key}{your user key generated from https://developer.meltwater.com}
}
\description{
access_token() generates a Bearer token, which has to be renewed every hour.
}
