% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_documents.R
\name{search_documents}
\alias{search_documents}
\title{Search documents}
\usage{
search_documents(keyword, start_date, end_date, type = "news")
}
\arguments{
\item{keyword}{the keyword you want to explore}

\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{type}{"news" or "social", defaults to "news"}
}
\description{
search_documents() enables you to search every document containg a keyword during a specific time interval
}
\examples{
df <- search_documents(user_key = user_key, token = token,
keyword = "demoskop",
start_date = "2017-07-01",
end_date = "2017-07-13",
type = "news_article")

}
