% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_documents.R
\name{search_documents}
\alias{search_documents}
\title{Search documents}
\usage{
search_documents(user_key, token, keyword, start_date, end_date, type = "all")
}
\arguments{
\item{user_key}{your user key generated from https://developer.meltwater.com}

\item{token}{your token generated from access_token()}

\item{keyword}{the keyword you want to explore}

\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{type}{"news_article", "social_post" or "all", defaults to "all"}
}
\description{
search_documents() enables you to search every document containg a keyword during a specific time interval
}
