% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_interval.R
\name{search_interval}
\alias{search_interval}
\title{Search counts of keyword during time interval}
\usage{
search_interval(keyword, start_date, end_date, type = "news",
  granularity = "DAY")
}
\arguments{
\item{keyword}{the keyword you want to explore, e.g. "iphone"}

\item{start_date}{start date, in "1900-01-01" format}

\item{end_date}{end date, in "1900-01-01" format}

\item{type}{"news" or "social", defaults to "news"}

\item{granularity}{defaults to "DAY" but "HOUR", "DAY", "WEEK" and "MONTH" are possible.}
}
\description{
search_interval() generates a data frame containing the number of articles or social posts written about your keyword during an interval of time
}
\examples{
df <- search_interval(user_key = user_key, token = token,
keyword = "demoskop",
start_date = "2017-07-01",
end_date = "2017-07-13",
type = "news_article",
granularity = "DAY")
}
